+++ Code/Backend/Backend/Repositories/IProductRepository.cs
@@ -6,7 +6,7 @@ public interface IProductRepository
 {
     Task<Product?> GetByIdAsync(long productId);
     Task<List<Product>> GetAllAsync();
-    Task<List<Product>> GetByCategoryAsync(int categoryId);
+    Task<List<Product>> GetByCategoryAsync(long categoryId);
     Task<List<Product>> GetByUserIdAsync(int userId);
     Task<List<Product>> GetAvailableAsync();
     Task AddAsync(Product product);
