+++ Code/Backend/Backend/Services/ICategoryService.cs
@@ -8,7 +8,7 @@ public interface ICategoryService
     ///<summary>
     ///根据Id获取分类信息
     ///</summary>
-    Task<CategoryDto?> GetByIdAsync(int categoryId);
+    Task<CategoryDto?> GetByIdAsync(long categoryId);
     ///<summary>
     ///获取所有分类
     ///</summary>
@@ -16,7 +16,7 @@ public interface ICategoryService
     ///<summary>
     ///获取子分类
     ///</summary>
-    Task<List<CategoryDto>> GetChildrenAsync(int parentId);
+    Task<List<CategoryDto>> GetChildrenAsync(long parentId);
     //TODO：以下考虑正式版本删除
    // Task<List<CategoryDto>> GetRootCategoriesAsync(); 获取根目录，可以通过GetAll加过滤获得
     ///<summary>
@@ -26,9 +26,9 @@ public interface ICategoryService
     ///<summary>
     ///删除分类
     ///</summary>
-    Task<bool> DeleteCategoryAsync(int categoryId);  //同理
+    Task<bool> DeleteCategoryAsync(long categoryId);  //同理
     ///<summary>
     ///获取某分类下的商品列表
     ///</summary>
-    Task<List<ProductDto>> GetProductsByCategoryAsync(int categoryId);//查某分类下的商品列表（占位，后续切到 ProductService,也可以不切，有待讨论）
+    Task<List<ProductDto>> GetProductsByCategoryAsync(long categoryId);//查某分类下的商品列表（占位，后续切到 ProductService,也可以不切，有待讨论）
 }
